"Fixed" collider scaling
Some colliders were giving warnings in the console, claiming they had negative scales. They didn't, but some of the houses had been flipped using a scale of -1 on one axis. This meant that the positive scaling of the collider meant it was technically negative. I've made these scales negative to account for this. It doesn't actually change anything in-game, but now the console isn't full of warnings :P