Fixed tweening easing
The loop to continue tweening was checking if the transform position was equal to the target position. If they didn't exactly match, the loop would keep running. This meant that the progress of the tween could go higher than 1. Given that the easing functions make use of sine and cosine, this resulted in the position of the transform going backwards and forwards as the progress went between integers. I've replaced this with a check of whether progress has reached 1 yet or not, as well as clamping the progress to 1. This should ensure that the tween will always end at a progress of exactly 1, and in the exact correct position.