Created method to cycle turn order UI
The method is async and returns the task so can be awaited. The method itself moves the first turn icon (the combatant who's turn has just finished) offscreen to the left, before shuffling the other icons down to the left and moving the first combatant to the back of the queue from offscreen to the right.